Fix problems with handling symlinks.
The old behavior was as follow.  If a symlink changed; or a file became
a symlink or the other way around, then pachy would try to create a
xdelta3 between them anyway, which usually fails.

The new behavior is more sane.  If a symlink changes value or a file
changes to a symlink or the other way around, then pachy will move the
old symlink/file to the deleted folder in the delta.

Signed-off-by: Bas Westerbaan <bas@westerbaan.name>